with sorting function
Figured out the alphabetizing mechanism without using util or scanner which is the most popular method.  Our instructor didn't want us to use it i'm guessing so we would learn what it does without having it done for us.  Just need to pretty this up a little and ready to submit!